Moving CSV::Row to their own file
    Implements part of https://github.com/ruby/csv/issues/5

    This does not change any behavior of the csv gem.
